[refactor] 방 정보 컴포넌트들을 room-info 폴더로 이동
